requirements.txt
paho-mqtt==1.4.0
redis==2.10.6
